서버 3주차
서버 3주차

+++ JSP_Project/WebContent/views/board/boardDetailView.jsp
@@ -1,8 +1,9 @@
-<%@ page import="com.kh.board.model.vo.Board" %>
+<%@ page import="com.kh.board.model.vo.*" %>
 <%@ page language="java" contentType="text/html; charset=UTF-8"
     pageEncoding="UTF-8"%>
 <%
 	Board b = (Board) request.getAttribute("b");
+	Attachment at = (Attachment) request.getAttribute("at");
 %>
 <!DOCTYPE html>
 <html>
@@ -45,22 +46,26 @@
 			<tr>
 				<th>첨부파일</th>
 				<td colspan="3">
-						
-					<a download="<%= b.getTitleImg() %>" href="<%= contextPath %>/resource/board_upfiles<%= b.getTitleImg() %>"><%= b.getTitleImg() %></a>
+					<% if(at == null) { %>
+						첨부파일이 없습니다
+					<% } else { %>
+						<!-- href='/jspproject/resources/board_upfiles/2022xxxx.jpg' -->
+						<a download="<%= at.getOriginName() %>" href="<%= contextPath %><%= at.getFilePath()+at.getChangeName() %>"><%= at.getOriginName() %></a>
+					<% } %>	
 				</td>
 			</tr>
 		</table>
-		
+
 		<br>
 		
 		<div align="center">
 			<a href="<%= contextPath %>/list.bo?currentPage=1" class="btn btn-secondary btn=sm">목록가기</a>
 			
 			<!-- 로그인한 사용자가 해당 게시글의 작성자인 경우 -->
-			<%-- <% if(liginUser != null && loginUser.getUserId().equals(b.getBoardWriter())) { %>
-				<a href="<%= contextPath %>/updateForm.bo?bno=<%= b.getBoardNo() %>" class="btn btn-warning btn-sm">수정하기</a>
+			<% if(loginUser != null && loginUser.getUserId().equals(b.getBoardWriter())) { %>
+				<a href="<%= contextPath %>/update.bo?bno=<%= b.getBoardNo() %>" class="btn btn-warning btn-sm">수정하기</a>
 				<a href="<%= contextPath %>/delete.bo?bno=<%= b.getBoardNo() %>" class="btn btn-danger btn-sm">삭제하기</a>
-			<% } %> --%>
+			<% } %>
 		</div>
 		
 		<br>
